+++ packages/generator-common/src/compiler.spec.ts
@@ -17,7 +17,7 @@ import {
 
 describe('compiler options', () => {
   const pathRootNodeModules = resolve(__dirname, '../../../node_modules');
-  beforeAll(() => {
+  beforeEach(() => {
     mock({
       [pathRootNodeModules]: mock.load(pathRootNodeModules),
       'config1/tsconfig.json': JSON.stringify({
@@ -42,7 +42,7 @@ describe('compiler options', () => {
     });
   });
 
-  afterAll(() => {
+  afterEach(() => {
     mock.restore();
   });
 
@@ -103,7 +103,7 @@ describe('compiler options', () => {
 });
 
 describe('compilation', () => {
-  beforeAll(async () => {
+  beforeEach(async () => {
     const rootNodeModules = resolve(__dirname, '../../../node_modules');
     const packageNodeModules = resolve(__dirname, '../node_modules');
     mock({
@@ -130,7 +130,7 @@ describe('compilation', () => {
     };
   }
 
-  afterAll(() => {
+  afterEach(() => {
     mock.restore();
   });
 
@@ -138,16 +138,27 @@ describe('compilation', () => {
     await transpileDirectory('test-src', compilerConfig('test-dist'));
     const files = await promises.readdir('test-dist');
     expect(files.includes('test-file.spec.js')).toBe(false);
-    expect(files).toIncludeAnyMembers([
-      'file-1.js',
+    expect(files).toEqual([
       'file-1.d.ts',
-      'file-1.d.ts.map'
+      'file-1.d.ts.map',
+      'file-1.js',
+      'file-1.js.map',
+      'index.d.ts',
+      'index.d.ts.map',
+      'index.js',
+      'index.js.map',
+      'sub-folder'
     ]);
     const filesSubfolder = await promises.readdir('test-dist/sub-folder');
-    expect(filesSubfolder).toIncludeAnyMembers([
-      'file-2.js',
+    expect(filesSubfolder).toEqual([
       'file-2.d.ts',
-      'file-2.d.ts.map'
+      'file-2.d.ts.map',
+      'file-2.js',
+      'file-2.js.map',
+      'index.d.ts',
+      'index.d.ts.map',
+      'index.js',
+      'index.js.map'
     ]);
   });
 
@@ -159,7 +170,7 @@ describe('compilation', () => {
     const files = new GlobSync('**/*.js', {
       cwd: 'test-dist-1'
     }).found;
-    expect(files).toIncludeSameMembers(['file-1.js', 'sub-folder/file-2.js']);
+    expect(files).toEqual(['file-1.js', 'sub-folder/file-2.js']);
   });
 
   it('considers exclude correctly', async () => {
@@ -170,7 +181,7 @@ describe('compilation', () => {
     const files = new GlobSync('**/*.js', {
       cwd: 'test-dist-2'
     }).found;
-    expect(files).toIncludeSameMembers(['file-1.js', 'sub-folder/file-2.js']);
+    expect(files).toEqual(['file-1.js', 'sub-folder/file-2.js']);
   });
 
   it('throws error with file information on broken source file', async () => {