+++ packages/mail-client/src/index.ts
@@ -1,3 +1,9 @@
+/**
+ * [[include:mail-client/README.md]]
+ * @packageDocumentation
+ * @module @sap-cloud-sdk/mail-client
+ */
+
 export { sendMail } from './mail-client';
 export type {
   MailConfig,
@@ -8,5 +14,7 @@ export type {
   Envelope,
   MailResponse,
   MailDestination,
-  MailClientOptions
+  MailClientOptions,
+  SmtpTransportOptions,
+  SDKOptions
 } from './mail-client-types';