+++ packages/openapi-generator/src/generator.ts
@@ -12,7 +12,6 @@ import {
   getSdkMetadataFileNames,
   getSdkVersion,
   readCompilerOptions,
-  sdkMetadataHeader,
   transpileDirectory,
   copyFiles,
   packageDescription
@@ -71,8 +70,11 @@ export async function generateWithParsedOptions(
   }
 
   if (options.clearOutputDir) {
+    // function rm was added in node version 14 and is the preferred method to use.
     const rm = promisesFs.rm || promisesFs.rmdir;
-    await rm(options.outputDir, { recursive: true });
+    const forceOption =
+      typeof promisesFs.rm === 'undefined' ? {} : { force: true };
+    await rm(options.outputDir, { recursive: true, ...forceOption });
   }
   const inputFilePaths = await getInputFilePaths(options.input);
 
@@ -303,26 +305,13 @@ function generateReadme(
 async function generateMetadata(
   openApiDocument: OpenApiDocument,
   inputFilePath: string,
-  { packageVersion, overwrite }: ParsedGeneratorOptions
+  { overwrite }: ParsedGeneratorOptions
 ) {
   const { name: inputFileName, dir: inputDirPath } = parse(inputFilePath);
-  const { clientFileName, headerFileName } =
-    getSdkMetadataFileNames(inputFileName);
+  const { clientFileName } = getSdkMetadataFileNames(inputFileName);
 
-  logger.verbose(`Generating header metadata ${headerFileName}.`);
   const metadataDir = resolve(inputDirPath, 'sdk-metadata');
   await mkdir(metadataDir, { recursive: true });
-  const headerFile = createFile(
-    metadataDir,
-    headerFileName,
-    JSON.stringify(
-      await sdkMetadataHeader('rest', inputFileName, packageVersion),
-      null,
-      2
-    ),
-    overwrite,
-    false
-  );
 
   logger.verbose(`Generating client metadata ${clientFileName}...`);
   const clientFile = createFile(
@@ -332,7 +321,7 @@ async function generateMetadata(
     overwrite,
     false
   );
-  return Promise.all([headerFile, clientFile]);
+  return clientFile;
 }
 
 async function generatePackageJson(