+++ test-packages/type-tests/test/jwt.test-d.ts
@@ -1,0 +1,4 @@
+import { decodeJwt, JwtPayload } from '@sap-cloud-sdk/connectivity';
+import { expectType } from 'tsd';
+
+expectType<JwtPayload>(decodeJwt(''));